refactor: allow all network configurations in scaffold.config
